1.0.2 | Save/Load function
Added a new function that allows you to save encrypted messages in a file that can later be opened. When the file is opened the program automatically discovers in which method the message is encrypted and decrypts it. A function that facilitates and speeds up the exchange of messages between users because all the decryption process takes place automatically.